tests/wpt/tests/html/browsers/windows/browsing-context-names/choose-_parent-003.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - '_parent' (via window.open)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  var topWindow;
  t.add_cleanup(() => topWindow.close());
  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data.name, 'parentTopReplace');
    assert_equals(e.data.isTop, true);
  }));
  topWindow = window.open('resources/choose-_parent-003-window.html', 'parentTopReplace');
}, '_parent should reuse window.parent context');
</script>
